digipeli/pyproject.toml
[tool.poetry]
name = "digipeli"
version = "0.1.0"
description = "Digitaalinen versio sananselityspelistä"
authors = ["Teemu Pennanen <teemu.pennanen1@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
invoke = "^1.7.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
coverage = "^6.3.2"
pytest-dotenv = "^0.5.2"
pylint = "^2.13.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
